Mode 1 + Mode 2: clean two-mode design, both verified end-to-end
Replaces the messy "Path A / Path B / vio= lio= toggle args" model with
a single-purpose launch file per concern. Two ways to use the system:

  Mode 1 — one command, full stack:
    docker compose up
    (cave.launch.py IncludeLaunchDescription's world + spawn + vio + lio)

  Mode 2 — step by step, four terminals:
    1. ros2 launch explorer_r2_sim world.launch.py
    2. ros2 launch explorer_r2_sim spawn_robot.launch.py
    3. ros2 launch explorer_r2_sim vio.launch.py
    4. ros2 launch explorer_r2_sim lio.launch.py

Files:
  - launch/world.launch.py (NEW): gz + world + bridge + RViz + joy.
    No robot, no estimators. Owns the world-preset resolver and the
    GZ_SIM_RESOURCE_PATH env setup.
  - launch/spawn_robot.launch.py (NEW): ros_gz_sim create the rover.
    Owns the SPAWN_POSES table and spawn_x/y/z/yaw args. Races safely
    against gz startup because ros_gz_sim create polls the spawn
    service.
  - launch/cave.launch.py (REFACTOR): thin Mode-1 wrapper that just
    IncludeLaunchDescription's world + spawn_robot + vio + lio.
    Forwards world / gui / rviz / joy / teleop / rqt_steering / verbose
    / spawn_{x,y,z,yaw} args. The previous vio/lio toggle args and
    the embedded gz_sim + spawn + bridge + RViz + joy code paths are
    gone; everything is composed from the four sub-launches.
  - launch/vio.launch.py / lio.launch.py: arg names stay
    `vio_config_path` / `lio_config_path` so cave.launch.py can include
    both without launch-arg collision (the previous `config_path`
    collision caused FAST_LIO to load OpenVINS YAML and crash).
  - scripts/lidar_field_adapter.py: emit `ring` as uint8 to match
    FAST_LIO's ouster_ros::Point definition (was uint16, PCL refused
    to bind).

README rewritten around Mode 1 + Mode 2. The old "Run VIO / Run LIO"
standalone sections are gone — they're now Steps 3 and 4 of Mode 2.

Verified both modes end-to-end with `docker compose run --rm sim
bash -ic "timeout 25 ros2 launch ... gui:=false rviz:=false joy:=false"`:
Mode 1 launches all 6 processes clean (gazebo, bridge, create, OpenVINS,
adapter, fastlio_mapping); each Mode 2 launch file starts clean on its
own. VIO reports "failed static init: no accel jerk" (correct,
stationary robot) and LIO no longer floods "No point, skip this scan".

+++ src/explorer_r2_sim/scripts/lidar_field_adapter.py
@@ -66,6 +66,10 @@ class LidarFieldAdapter(Node):
     # ------------------------------------------------------------------
     # Output point dtype matches ouster_ros::Point field names — FAST_LIO
     # looks these up by name via PCL's PointCloud2 → PointCloud conversion.
+    # Field types match FAST_LIO's ouster_ros::Point definition exactly
+    # (preprocess.h: ring is std::uint8_t, not uint16). PCL matches
+    # PointCloud2 fields by NAME and TYPE — a UINT16 ring won't bind to
+    # a uint8_t struct member.
     _OUT_DTYPE = np.dtype([
         ("x",            np.float32),
         ("y",            np.float32),
@@ -73,7 +77,7 @@ class LidarFieldAdapter(Node):
         ("intensity",    np.float32),
         ("t",            np.uint32),
         ("reflectivity", np.uint16),
-        ("ring",         np.uint16),
+        ("ring",         np.uint8),
         ("ambient",      np.uint16),
         ("range",        np.uint32),
     ])
@@ -85,7 +89,7 @@ class LidarFieldAdapter(Node):
         ("intensity",    PointField.FLOAT32),
         ("t",            PointField.UINT32),
         ("reflectivity", PointField.UINT16),
-        ("ring",         PointField.UINT16),
+        ("ring",         PointField.UINT8),
         ("ambient",      PointField.UINT16),
         ("range",        PointField.UINT32),
     ]
@@ -106,7 +110,7 @@ class LidarFieldAdapter(Node):
         r_xy = np.sqrt(x * x + y * y)
         elev = np.arctan2(z, np.maximum(r_xy, 1e-9))
         ring_f = (elev - self.min_elev) / self.elev_span * (self.N - 1)
-        ring = np.clip(np.round(ring_f), 0, self.N - 1).astype(np.uint16)
+        ring = np.clip(np.round(ring_f), 0, self.N - 1).astype(np.uint8)
 
         # t (per-point time offset within scan) — linear in azimuth
         azi = np.arctan2(y, x)
